requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
executing==0.5.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
horovod==0.20.3
idna==2.10
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3
matplotlib==3.3.2
mpi4py==3.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
openslide-python==1.1.2
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.50.1
typeguard==2.9.1
urllib3==1.25.10
varname==0.4.0
Werkzeug==1.0.1
wrapt==1.12.1
